start @blitzjs/core/server entrypoint
