+++ packages/blitz/src/index.ts
@@ -4,3 +4,4 @@ export * from "@blitzjs/core/dynamic"
 export * from "@blitzjs/core/head"
 export * from "@blitzjs/core/image"
 export * from "@blitzjs/core"
+export * from "@blitzjs/core/server"
