+++ packages/core/src/dynamic.ts
@@ -1,1 +1,6 @@
+/*
+ * IF YOU CHANGE THIS FILE
+ *    You also need to update the rewrite map in
+ *    packages/babel-preset/src/rewrite-imports.ts
+ */
 export {default as dynamic, noSSR} from "next/dynamic"
