+++ packages/core/src/head.ts
@@ -1,3 +1,8 @@
+/*
+ * IF YOU CHANGE THIS FILE
+ *    You also need to update the rewrite map in
+ *    packages/babel-preset/src/rewrite-imports.ts
+ */
 import {default as NextHead} from "next/head"
 export * from "next/head"
 export const Head = NextHead
