+++ packages/core/src/image.ts
@@ -1,3 +1,8 @@
+/*
+ * IF YOU CHANGE THIS FILE
+ *    You also need to update the rewrite map in
+ *    packages/babel-preset/src/rewrite-imports.ts
+ */
 import {default as NextImage} from "next/image"
 export * from "next/image"
 export const Image = NextImage
