+++ packages/core/src/index.ts
@@ -7,21 +7,17 @@ export * from "./constants"
 export {withBlitzAppRoot} from "./blitz-app-root"
 export {useQuery, usePaginatedQuery, useInfiniteQuery} from "./use-query-hooks"
 export {getQueryKey, invalidateQuery, setQueryData} from "./utils/react-query-utils"
-export {passportAuth} from "./passport-adapter"
+export {passportAuth} from "./server/passport-adapter"
 export {getIsomorphicEnhancedResolver} from "./rpc"
 export {useMutation} from "./use-mutation"
-export {invoke, invokeWithMiddleware} from "./invoke"
+export {invoke} from "./invoke"
 export {getBlitzRuntimeData} from "./blitz-data"
 export {resolver} from "./resolver"
 export type {AuthenticatedMiddlewareCtx} from "./resolver"
 export {paginate} from "./server-utils"
 
-export {
-  getAllMiddlewareForModule,
-  handleRequestWithMiddleware,
-  connectMiddleware,
-} from "./middleware"
-export type {Ctx, DefaultCtx} from "./middleware"
+export {SecurePassword, hash256, generateToken} from "./auth-utils"
+
 export {
   getAntiCSRFToken,
   useSession,
@@ -36,8 +32,6 @@ export type {
   ClientSession,
   AuthenticatedClientSession,
 } from "./supertokens"
-
-export {SecurePassword, hash256, generateToken} from "./auth-utils"
 
 export {isLocalhost} from "./utils/index"
 export {prettyMs} from "./utils/pretty-ms"
