+++ packages/core/src/server/index.ts
@@ -1,0 +1,12 @@
+/*
+ * IF YOU CHANGE THIS FILE
+ *    You also need to update the rewrite map in
+ *    packages/babel-preset/src/rewrite-imports.ts
+ */
+export {
+  getAllMiddlewareForModule,
+  handleRequestWithMiddleware,
+  connectMiddleware,
+} from "./middleware"
+
+export {invokeWithMiddleware} from "./invoke-with-middleware"
