+++ packages/core/src/supertokens.ts
@@ -2,9 +2,8 @@ import {useEffect, useState} from "react"
 import {getBlitzRuntimeData} from "./blitz-data"
 import {COOKIE_CSRF_TOKEN, COOKIE_LEGACY_CSRF_TOKEN} from "./constants"
 import {AuthenticationError} from "./errors"
-import {Ctx} from "./middleware"
 import {publicDataStore} from "./public-data-store"
-import {EmptyPublicData, IsAuthorizedArgs, PublicData} from "./types"
+import {Ctx, EmptyPublicData, IsAuthorizedArgs, PublicData} from "./types"
 import {isServer} from "./utils"
 import {readCookie} from "./utils/cookie"
